Use gnu tar to build the docsets
Previously, we used `:erl_tar`, the inbuilt tar library which comes with
OTP. However, we found that Zeal needs and expects directories to be
defined in the .tgz files, and `:erl_tar` is not capable of including
directories. Therefore, in order to make .tgz files which are compatible
with Zeal, we now use tar to build the archive, which does include the
directories.